+++ src/accessories/filterMaintenancePlatformAccessory.ts
@@ -3,6 +3,7 @@ import { AccessoryPlugin, API, Service, CharacteristicValue } from 'homebridge';
 import { QueueSendItem, QueueReceiveItem } from "../queue";
 import { ErrorNumber } from "../error";
 import { md5 } from "../md5";
+import { UdpClient } from '../udp';
 
 export class FilterMaintenancePlatformAccessory implements AccessoryPlugin {
 
@@ -15,8 +16,11 @@ export class FilterMaintenancePlatformAccessory implements AccessoryPlugin {
   private platform: any;
   private device: any;
   private pushButton: number;
+  private logging: number;
   private updateFilterChangeIndicationQueued: boolean;
   private updateFilterLifeLevelQueued: boolean;
+
+  private udpClient: UdpClient;
 
   private accStates = {
     FilterChangeIndication: 0,
@@ -33,6 +37,9 @@ export class FilterMaintenancePlatformAccessory implements AccessoryPlugin {
     this.platform   = platform;
     this.device     = device;
     this.pushButton = this.device.pushButton || this.platform.pushButton;
+    this.logging    = this.device.logging    || 0;
+
+    this.udpClient = new UdpClient(this.platform, this.device);
 
     this.errorCheck();
 
@@ -125,6 +132,10 @@ export class FilterMaintenancePlatformAccessory implements AccessoryPlugin {
         }
 
         this.service.updateCharacteristic(this.api.hap.Characteristic.FilterChangeIndication, this.accStates.FilterChangeIndication);
+
+        if (this.logging) {
+          this.udpClient.sendMessage("FilterChangeIndication", String(this.accStates.FilterChangeIndication));
+        }
       }
 
       this.updateFilterChangeIndicationQueued = false;
@@ -154,6 +165,10 @@ export class FilterMaintenancePlatformAccessory implements AccessoryPlugin {
           }
   
           this.service.updateCharacteristic(this.api.hap.Characteristic.FilterLifeLevel, this.accStates.FilterLifeLevel);
+
+          if (this.logging) {
+            this.udpClient.sendMessage("FilterLifeLevel", String(this.accStates.FilterLifeLevel));
+          }
         }
 
         this.updateFilterLifeLevelQueued = false;