+++ src/accessories/irrigationSystemPlatformAccessory.ts
@@ -3,6 +3,7 @@ import { AccessoryPlugin, API, Service, CharacteristicValue } from 'homebridge';
 import { QueueSendItem, QueueReceiveItem } from "../queue";
 import { ErrorNumber } from "../error";
 import { md5 } from "../md5";
+import { UdpClient } from '../udp';
 import { ValvePlatformAccessory } from './valvePlatformAccessory';
 
 
@@ -20,12 +21,15 @@ export class IrrigationSystemPlatformAccessory implements AccessoryPlugin {
   private platform: any;
   private device: any;
   private pushButton: number;
+  private logging: number;
   private irrigationSystemAutoUpdate: number;
   private updateActiveQueued: boolean;
   private updateProgramModeQueued: boolean;
   private updateInUseQueued: boolean;
   private updateWaterLevelQueued: boolean;
   private updateRemainingDurationQueued: boolean;
+
+  private udpClient: UdpClient;
 
   private accStates = {
     Active: 0,
@@ -44,6 +48,10 @@ export class IrrigationSystemPlatformAccessory implements AccessoryPlugin {
     this.platform   = platform;
     this.device     = device;
     this.pushButton = this.device.pushButton || this.platform.pushButton;
+    this.logging    = this.device.logging    || 0;
+
+    this.udpClient = new UdpClient(this.platform, this.device);
+
     this.irrigationSystemAutoUpdate = (this.device.irrigationSystemAutoUpdate ? 1 : 0);
     this.valveAccessories = [];
     this.services = [];
@@ -215,6 +223,10 @@ export class IrrigationSystemPlatformAccessory implements AccessoryPlugin {
           }
 
           this.service.updateCharacteristic(this.api.hap.Characteristic.Active, this.accStates.Active);
+
+          if (this.logging) {
+            this.udpClient.sendMessage("Active", String(this.accStates.Active));
+          }
         }
 
         this.updateActiveQueued = false;
@@ -243,6 +255,10 @@ export class IrrigationSystemPlatformAccessory implements AccessoryPlugin {
         }
 
         this.service.updateCharacteristic(this.api.hap.Characteristic.ProgramMode, this.accStates.ProgramMode);
+
+        if (this.logging) {
+          this.udpClient.sendMessage("ProgramMode", String(this.accStates.ProgramMode));
+        }
       }
 
       this.updateProgramModeQueued = false;
@@ -282,6 +298,10 @@ export class IrrigationSystemPlatformAccessory implements AccessoryPlugin {
           }
 
           this.service.updateCharacteristic(this.api.hap.Characteristic.InUse, this.accStates.InUse);
+
+          if (this.logging) {
+            this.udpClient.sendMessage("InUse", String(this.accStates.InUse));
+          }
         }
 
         this.updateInUseQueued = false;
@@ -312,6 +332,10 @@ export class IrrigationSystemPlatformAccessory implements AccessoryPlugin {
           }
 
           this.service.updateCharacteristic(this.api.hap.Characteristic.RemainingDuration, this.accStates.RemainingDuration);
+
+          if (this.logging) {
+            this.udpClient.sendMessage("RemainingDuration", String(this.accStates.RemainingDuration));
+          }
 
         }
 
@@ -344,6 +368,10 @@ export class IrrigationSystemPlatformAccessory implements AccessoryPlugin {
           }
 
           this.service.updateCharacteristic(this.api.hap.Characteristic.WaterLevel, this.accStates.WaterLevel);
+
+          if (this.logging) {
+            this.udpClient.sendMessage("WaterLevel", String(this.accStates.WaterLevel));
+          }
         }
 
         this.updateWaterLevelQueued = false;