+++ src/accessories/switchPlatformAccessory.ts
@@ -6,6 +6,7 @@ import { AccessoryPlugin, API, Service, CharacteristicValue } from 'homebridge';
 import { QueueSendItem, QueueReceiveItem } from "../queue";
 import { ErrorNumber } from "../error";
 import { md5 } from "../md5";
+import { UdpClient } from '../udp';
 
 export class SwitchPlatformAccessory implements AccessoryPlugin {
 
@@ -18,7 +19,10 @@ export class SwitchPlatformAccessory implements AccessoryPlugin {
   private platform: any;
   private device: any;
   private pushButton: number;
+  private logging: number;
   private updateOnQueued: boolean;
+
+  private udpClient: UdpClient;
 
   private accStates = {
     On: false,
@@ -33,6 +37,9 @@ export class SwitchPlatformAccessory implements AccessoryPlugin {
     this.platform   = platform;
     this.device     = device;
     this.pushButton = this.device.pushButton || this.platform.pushButton;
+    this.logging    = this.device.logging    || 0;
+
+    this.udpClient = new UdpClient(this.platform, this.device);
 
     this.errorCheck();
 
@@ -108,10 +115,14 @@ export class SwitchPlatformAccessory implements AccessoryPlugin {
         this.accStates.On = on;
 
         if (this.platform.config.debugMsgLog || this.device.debugMsgLog) {
-          this.platform.log.info('[%s] Get On -> %s', this.device.name, on);
+          this.platform.log.info('[%s] Get On -> %s', this.device.name, this.accStates.On);
         }
 
-        this.service.updateCharacteristic(this.api.hap.Characteristic.On, on);
+        this.service.updateCharacteristic(this.api.hap.Characteristic.On, this.accStates.On);
+
+        if (this.logging) {
+          this.udpClient.sendMessage("On", String(this.accStates.On));
+        }
       }
 
       this.updateOnQueued = false;