+++ src/accessories/thermostatPlatformAccessory.ts
@@ -3,6 +3,7 @@ import { AccessoryPlugin, API, Service, CharacteristicValue } from 'homebridge';
 import { QueueSendItem, QueueReceiveItem } from "../queue";
 import { ErrorNumber } from "../error";
 import { md5 } from "../md5";
+import { UdpClient } from '../udp';
 
 export class ThermostatPlatformAccessory implements AccessoryPlugin {
 
@@ -15,10 +16,13 @@ export class ThermostatPlatformAccessory implements AccessoryPlugin {
   private platform: any;
   private device: any;
   private pushButton: number;
+  private logging: number;
   private updateCurrentHeatingCoolingStateQueued: boolean;
   private updateTargetHeatingCoolingStateQueued: boolean;
   private updateCurrentTemperatureQueued: boolean;
   private updateTargetTemperatureQueued: boolean;
+
+  private udpClient: UdpClient;
 
   private accStates = {
     CurrentHeatingCoolingState: 0,
@@ -39,6 +43,9 @@ export class ThermostatPlatformAccessory implements AccessoryPlugin {
     this.platform   = platform;
     this.device     = device;
     this.pushButton = this.device.pushButton || this.platform.pushButton;
+    this.logging    = this.device.logging    || 0;
+
+    this.udpClient = new UdpClient(this.platform, this.device);
 
     this.errorCheck();
 
@@ -183,6 +190,10 @@ export class ThermostatPlatformAccessory implements AccessoryPlugin {
         }
 
         this.service.updateCharacteristic(this.api.hap.Characteristic.CurrentHeatingCoolingState, this.accStates.CurrentHeatingCoolingState);
+
+        if (this.logging) {
+          this.udpClient.sendMessage("CurrentHeatingCoolingState", String(this.accStates.CurrentHeatingCoolingState));
+        }
       }
 
       this.updateCurrentHeatingCoolingStateQueued = false;
@@ -210,6 +221,10 @@ export class ThermostatPlatformAccessory implements AccessoryPlugin {
         }
 
         this.service.updateCharacteristic(this.api.hap.Characteristic.TargetHeatingCoolingState, this.accStates.TargetHeatingCoolingState);
+
+        if (this.logging) {
+          this.udpClient.sendMessage("TargetHeatingCoolingState", String(this.accStates.TargetHeatingCoolingState));
+        }
       }
 
       this.updateTargetHeatingCoolingStateQueued = false;
@@ -247,6 +262,10 @@ export class ThermostatPlatformAccessory implements AccessoryPlugin {
         }
 
         this.service.updateCharacteristic(this.api.hap.Characteristic.CurrentTemperature, this.accStates.CurrentTemperature);
+
+        if (this.logging) {
+          this.udpClient.sendMessage("CurrentTemperature", String(this.accStates.CurrentTemperature));
+        }
       }
 
       this.updateCurrentTemperatureQueued = false;
@@ -285,6 +304,10 @@ export class ThermostatPlatformAccessory implements AccessoryPlugin {
         }
 
         this.service.updateCharacteristic(this.api.hap.Characteristic.TargetTemperature, this.accStates.TargetTemperature);
+
+        if (this.logging) {
+          this.udpClient.sendMessage("TargetTemperature", String(this.accStates.TargetTemperature));
+        }
       }
 
       this.updateTargetTemperatureQueued = false;