+++ src/sensors/contactSensorPlatformAccessory.ts
@@ -3,6 +3,7 @@ import { AccessoryPlugin, API, Service, CharacteristicValue } from 'homebridge';
 import { QueueReceiveItem } from "../queue";
 import { ErrorNumber } from "../error";
 import { md5 } from "../md5";
+import { UdpClient } from '../udp';
 
 export class ContactSensorPlatformAccessory implements AccessoryPlugin {
 
@@ -14,7 +15,10 @@ export class ContactSensorPlatformAccessory implements AccessoryPlugin {
 
   private platform: any;
   private device: any;
+  private logging: number;
   private updateContactSensorStateQueued: boolean;
+
+  private udpClient: UdpClient;
 
   private sensStates = {
     ContactSensorState: 0,
@@ -28,6 +32,9 @@ export class ContactSensorPlatformAccessory implements AccessoryPlugin {
     this.api      = api;
     this.platform = platform;
     this.device   = device;
+    this.logging  = this.device.logging || 0;
+
+    this.udpClient = new UdpClient(this.platform, this.device);
 
     this.errorCheck();
 
@@ -87,6 +94,10 @@ export class ContactSensorPlatformAccessory implements AccessoryPlugin {
         }
 
         this.service.updateCharacteristic(this.api.hap.Characteristic.ContactSensorState, this.sensStates.ContactSensorState);
+
+        if (this.logging) {
+          this.udpClient.sendMessage("ContactSensorState", String(this.sensStates.ContactSensorState));
+        }
       }
 
       this.updateContactSensorStateQueued = false;